+++ engine/runtime/geometry/src/csg/mod.rs
@@ -32,3 +32,4 @@ pub enum Csg {
 
 // Tessellation is defined only for `Polygon` / `PolygonMesh`.
 crate::unsupported!(Csg: Triangulate);
+crate::unsupported!(Csg: Reproject);
