+++ engine/runtime/geometry/src/line_string/ops.rs
@@ -1,5 +1,7 @@
 use super::{LineString2D, LineString3D};
-use crate::ops::{Aabb, BoundingBox, UnsupportedOperation};
+use crate::coordinate::{Coordinate, EpsgCode};
+use crate::ops::reproject::{transform_coords_2d, transform_coords_3d};
+use crate::ops::{Aabb, BoundingBox, Reproject, ReprojectionCache, UnsupportedOperation};
 
 impl BoundingBox for LineString2D {
     fn bounding_box(&self) -> Result<Aabb, UnsupportedOperation> {
@@ -17,6 +19,36 @@ impl BoundingBox for LineString3D {
             geometry: "LineString3D",
             operation: "bounding_box",
         })
+    }
+}
+
+impl Reproject for LineString2D {
+    fn reproject(
+        &mut self,
+        target: EpsgCode,
+        cache: &mut ReprojectionCache,
+    ) -> crate::error::Result<()> {
+        let from = self.coordinate.require_crs()?;
+        if from != target {
+            transform_coords_2d(cache, from, target, &mut self.coords, self.z.as_deref_mut())?;
+            self.coordinate = Coordinate::Crs(target);
+        }
+        Ok(())
+    }
+}
+
+impl Reproject for LineString3D {
+    fn reproject(
+        &mut self,
+        target: EpsgCode,
+        cache: &mut ReprojectionCache,
+    ) -> crate::error::Result<()> {
+        let from = self.coordinate.require_crs()?;
+        if from != target {
+            transform_coords_3d(cache, from, target, &mut self.coords)?;
+            self.coordinate = Coordinate::Crs(target);
+        }
+        Ok(())
     }
 }
