+++ engine/runtime/geometry/src/point/ops.rs
@@ -1,5 +1,6 @@
 use super::{Point2D, Point3D};
-use crate::ops::{Aabb, BoundingBox, UnsupportedOperation};
+use crate::coordinate::{Coordinate, EpsgCode};
+use crate::ops::{Aabb, BoundingBox, Reproject, ReprojectionCache, UnsupportedOperation};
 
 impl BoundingBox for Point2D {
     fn bounding_box(&self) -> Result<Aabb, UnsupportedOperation> {
@@ -10,6 +11,38 @@ impl BoundingBox for Point2D {
 impl BoundingBox for Point3D {
     fn bounding_box(&self) -> Result<Aabb, UnsupportedOperation> {
         Ok(Aabb::point_3d(self.position))
+    }
+}
+
+impl Reproject for Point2D {
+    fn reproject(
+        &mut self,
+        target: EpsgCode,
+        cache: &mut ReprojectionCache,
+    ) -> crate::error::Result<()> {
+        let from = self.coordinate.require_crs()?;
+        if from != target {
+            let [x, y] = self.position;
+            let [nx, ny, _] = cache.transform(from, target, [x, y, 0.0])?;
+            self.position = [nx, ny];
+            self.coordinate = Coordinate::Crs(target);
+        }
+        Ok(())
+    }
+}
+
+impl Reproject for Point3D {
+    fn reproject(
+        &mut self,
+        target: EpsgCode,
+        cache: &mut ReprojectionCache,
+    ) -> crate::error::Result<()> {
+        let from = self.coordinate.require_crs()?;
+        if from != target {
+            self.position = cache.transform(from, target, self.position)?;
+            self.coordinate = Coordinate::Crs(target);
+        }
+        Ok(())
     }
 }
