+++ engine/runtime/geometry/src/solid/mod.rs
@@ -8,12 +8,9 @@
 //! CityGML `Surface` name.) Solids are 3D only; their shells are coordless
 //! raw meshes and the one frame lives on the `Solid`.
 
-use crate::coordinate::EpsgCode;
 use serde::{Deserialize, Serialize};
 
 use crate::coordinate::Coordinate;
-use crate::error::Result;
-use crate::ops::reproject::{transform_coords_3d, ReprojectionCache};
 use crate::polygon_mesh::PolygonMesh3DData;
 use crate::triangular_mesh::TriangularMesh3DData;
 
@@ -50,37 +47,4 @@ pub struct Solid {
     exterior: Shell,
     /// Hollow voids.
     interiors: Vec<Shell>,
-}
-
-impl Solid {
-    /// Reproject all shell vertices to `target` (EPSG).
-    pub(crate) fn reproject(
-        &mut self,
-        target: EpsgCode,
-        cache: &mut ReprojectionCache,
-    ) -> Result<()> {
-        let from = self.coordinate.require_crs()?;
-        if from != target {
-            reproject_shell(&mut self.exterior, from, target, cache)?;
-            for shell in &mut self.interiors {
-                reproject_shell(shell, from, target, cache)?;
-            }
-            self.coordinate = Coordinate::Crs(target);
-        }
-        Ok(())
-    }
-}
-
-/// Reproject one shell's vertices from `from` to `target` (EPSG).
-fn reproject_shell(
-    shell: &mut Shell,
-    from: EpsgCode,
-    target: EpsgCode,
-    cache: &mut ReprojectionCache,
-) -> Result<()> {
-    let vertices = match shell {
-        Shell::PolygonMesh(data) => data.vertices_mut(),
-        Shell::TriangularMesh(data) => data.vertices_mut(),
-    };
-    transform_coords_3d(cache, from, target, vertices)
 }
